+++ conjure-python-core/src/main/java/com/palantir/conjure/python/types/PythonBeanGenerator.java
@@ -27,8 +27,6 @@ public interface PythonBeanGenerator {
     PythonSnippet generateType(
             TypeDefinition typeDef,
             Function<TypeName, ImportTypeVisitor> importTypeVisitorFactory,
-            DealiasingTypeVisitor dealiasingTypeVisitor,
-            TypeMapper typeMapper,
-            TypeMapper myPyMapper);
+            DealiasingTypeVisitor dealiasingTypeVisitor);
 
 }
