--- conjure-python-core/src/main/java/com/palantir/conjure/python/types/TypeMapper.java
@@ -1,33 +1,0 @@
-/*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.conjure.python.types;
-
-
-import com.palantir.conjure.spec.Type;
-
-public final class TypeMapper {
-
-    private final Type.Visitor<String> typeNameVisitor;
-
-    public TypeMapper(Type.Visitor<String> typeNameVisitor) {
-        this.typeNameVisitor = typeNameVisitor;
-    }
-
-    public String getTypeName(Type type) {
-        return type.accept(typeNameVisitor);
-    }
-}
